fix: correct parameter type (#1170)
### Briefly, what does this PR introduce?
Fixes a parameter type in the seeder to match the expected type in Acts.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #1167 )
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
No

+++ src/algorithms/tracking/OrthogonalTrackSeedingConfig.h
@@ -25,7 +25,7 @@ namespace eicrecon {
     float m_collisionRegionMin = -250 * Acts::UnitConstants::mm; // Min z for primary vertex
     float m_collisionRegionMax = 250 * Acts::UnitConstants::mm; // Max z for primary vertex
 
-    float m_maxSeedsPerSpM = 0; // max number of seeds a single middle sp can belong to - 1
+    unsigned int m_maxSeedsPerSpM = 0; // max number of seeds a single middle sp can belong to - 1
     float m_cotThetaMax = 1.0 / tan(2. * atan(exp(-4.0))); // Cotangent of max theta angle (based on eta)
 
     float m_sigmaScattering = 5; // How many standard devs of scattering angles to consider
